init move to scanners module
