+++ anubis/scanners/brute_force.py
@@ -1,0 +1,4 @@
+# TODO - Need to implement brute forcing in a non blocking, multi threaded way
+def brute_force(self, target):
+  print(
+    "NOT YET IMPLEMENTED - Starting brute force enumartin (warning - will take a while)")
